Change Evaluator evaluate to return a double for more precision

+++ app/src/main/java/scrabble/engine/Evaluator.java
@@ -3,5 +3,5 @@ package scrabble.engine;
 import scrabble.core.GameState;
 
 public interface Evaluator {
-    int evaluate(GameState gameState);
+    double evaluate(GameState gameState);
 }
